Added XVimEqualEvaluator for indent
Added support for:
'==', '={motion}', '{Visual}='

NOTE: It is a little buggy as I could not find a function I could call
to do the indentation, so I am faking a key press to do it for now.
Apparently NSResponder -indent: is declared but not implemented.

Cleaned up some warnings by exposing functions to public interface.