Fix for issue #51 and better approach to xvim.mode
Note that the caret is being drawn using drawRect: when in visual mode.
It turns out that it is almost impossible to do it any other way. I got
this information from the macvim source code. Here is a link to the
source and the comment itself:

http://macvim.googlecode.com/svn/trunk/MMTextView.m

NOTE: The insertion point is drawn manually in drawRect:.  It would be
nice to be able to use the insertion point related methods of
NSTextView, but it seems impossible to get them to work properly
(search the cocoa builder archives).