Fix problem with keymaps and modifiers.
An event is not processed differently depending on which modifier keys
are held down. Each event generates N number of possible key strokes
that are checked against the key map. If none are found, a primary key
stroke is selected and passed to the rest of the system.

For instance, assume shift-N is pressed. These key strokes are checked
against the key map:

    S-n
    S-N
    N (primary)

If C-f is pressed, these key strokes are checked against they key map:
    C-f (primary)