Introduce evaluator contexts.
An evaluator context encapsulates state about:
* Numeric argument
* Register argument
* Argument string (eg. "adi, shown in bottom right)

When an evaluator instantiates a new evaluator, it is the caller's
responsibility to create and set up the new context for that evaluator.

For example, if you create a delete evaluator you should pass the
current numeric argument, the register argument and append "d" to the
argument string. There are convencience functions that take care of most
of this for you.

Previously most of this state was global which is harder to reason about
and manage.