Improve encapsulation of XVimTextObjectEvaluator, improve clearing of the argument string
only have a "motion" selector to return the computed motion, instead of
everybody doing that.

Also improve the status line when the text evaluator is active in visual
mode: keep printing -- VISUAL* -- and do not forget to reset the argument
string if the text evaluator failed.

Do not fail the visual mode if the motion fails, or a g command fails,
which is what vim does (it also does a bell in that case, but oh well).