Reimplement xvim_shift to handle block shift properly
Do not use xcode shiftRight/shiftLeft anymore, it's easy to implement and
we need the work done for shifting at a given column, and what XCode does
is just the shift at column 0, so just use our own work.

rework xvim_blockInsertFixupWithText to operate on the first line too, and
adapt XVimInsertEvaluator to handle that change, because
xvim_blockInsertFixupWithText does the bulk of what ShiftRight is.

This implements `v_b_<` and `v_b_>`