Fix a few mistakes, and support movement |
When normal-mode replace failed, it used to fail to quit insert mode,
which broke several motions. Make -escapeFromInsert take a boolean to know
whether in addition to quitting INSERT mode, we should also go backwards.

Replace also had an off-by-one that prevented to replace the Last char of
the line (a >= comparison really should be >).

Add a BAR selector to the XVimMotionEvaluator, and a new
MOTION_COLUMN_OF_LINE enum value.