Fix Shift-O on the last line of the file
This fixes an issue where, when the start of the line is the last
character of the file, there an off by 1 error that would cause shift-o
not to work. `characterAtIndex` expects and index within the bounds of
this string, the passed character index was previously the length of the
string.

This fixes #675