Fix behavior of NUMBER/searchCurrentWordForward
Only search backwards an extra time if not at the beginning of a word.

The method `searchCurrentWordForward` when parameter `forward` is `NO`
previously skipped a word if the cursor is at the beginning of a word.
Repeatedly invoking `NUMBER` would skip every other word. NB when
searching backward (`QUESTION`) while in the middle of the searched
word, the first match is the word at the cursor. This was incorrectly
compensated for by searching backwards an extra time, regardless of
cursor location in the current word.