Added feature for blockcursor.
While we can't support everything vim can for guicursor, we can create our own proprietary option (similar to blinkcursor).
I decided to call this new option, blockcursor. It only controls the cursor in insert mode. I couldn't come up with a clever and concise option name that was more indicative of this.