Make unit tests whitespace agnostic.
Currently, if Xcode is set to indent with tabs a few unit tests (that are dependent upon whitespace settings) fail. We now read in the Xcode preferences and base our tests off of that setting.

Note: Although it is easy enough to detect tabs/spaces indentation and although it is easy enough to detect how many spaces are inserted for space indentation, it is not as straightforward to test spaces. The biggest problem is the tabWidth for inserting spaces. This can vary dramatically and since most of our test case strings are hard coded, this is not such an easy task. This makes tab indentation testing look more appealing, actually, in my humble opinion. If we wanted to support various tabWidth's for space indentation, we would need to start generating some of the unit test strings programmatically; something I'm just not quite willing to invest time in yet. Have other higher priorities ;)