Updates readme with link to XVim2
This is a link to XVim2 to let people know there is a working version, long term we might want to move the current `master` branch of XVim to `oldmaster` and migrate the code from XVim2 to XVim and make it the `master` branch on XVim. That repo has better SEO and if XVim2 is the future then it might make sense to just move it over.